Step5 [test] 규칙8.일급컬렉션
- 테스트 코드 수정
- Winner 클래스 매개변수를 List<Car>로 적용.
