+++ src/main/java/step4/view/ResultView.java
@@ -3,6 +3,7 @@ package step4.view;
 import step4.domain.Car;
 
 import java.util.List;
+import java.util.stream.Stream;
 
 public class ResultView {
 
@@ -15,9 +16,10 @@ public class ResultView {
         System.out.println(line.toString());
     }
 
-    public void renderWinner(List<String> winners) {
+    public void renderWinner(List<Car> winners) {
         if (winners.size() > 0) {
-            System.out.println(String.join(",", winners) + "가 최종 우승했습니다.");
+            Stream<Car> stream = winners.stream();
+            stream.forEach(s -> System.out.println(String.join(",", s.name()) + "가 최종 우승했습니다."));
         } else {
             System.out.println("우승자는 없습니다.");
         }
